Rebased the branch onto develop Moved tests into the right package Resolved conflicts
