+++ app/src/main/java/com/amalitech/bookmeetingroom/MainActivity.kt
@@ -10,6 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import com.amalitech.authentication.login.LoginScreen
 import com.example.core_ui.ui.theme.BookMeetingRoomTheme
 
 class MainActivity : ComponentActivity() {
@@ -22,7 +23,11 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Greeting("Android")
+                    LoginScreen(
+                        onNavigateToHome = { /*TODO*/ },
+                        onNavigateToForgotPassword = { /*TODO*/ },
+                        onNavigateToSignUp = { /*TODO*/ },
+                        onNavigateUp = { /*TODO*/ })
                 }
             }
         }
