+++ domain/authentication/src/main/java/com/example/authentication/use_case/LogIn.kt
@@ -13,7 +13,6 @@ class LogIn {
      */
     operator fun invoke(email: String, password: String): UiText? {
         // TODO(use the api to log into the account)
-        return null
         return UiText.DynamicString("The api is not yet available")
     }
 }
