+++ ui/authentication/build.gradle.kts
@@ -4,5 +4,8 @@ apply {
 
 dependencies {
     "implementation"(project(Modules.core))
+    "implementation"(project(Modules.coreUI))
     "implementation"(project(Modules.domainAuthentication))
+
+    "implementation"(Lifecycle.composeLifecycle)
 }
